refactor: ディレクトリ構造変更
