+++ app/enums/rich_text_type.py
@@ -1,10 +1,10 @@
 from enum import Enum
 from typing import Union
 
-from models.message_content import MessageContent
-from models.rich_text_list import RichTextList
-from models.rich_text_preformatted import RichTextPreformatted
-from models.rich_text_section import RichTextSection
+from app.models.message_content import MessageContent
+from app.models.rich_text_list import RichTextList
+from app.models.rich_text_preformatted import RichTextPreformatted
+from app.models.rich_text_section import RichTextSection
 
 
 class RichTextType(Enum):
